[TML] Inheritances introduced -> some functionalities moved to parent TreeModel
