+++ TreeModelLib/AbovegroundCompetition/__init__.py
@@ -7,3 +7,4 @@ Created on Thu Nov  8 15:25:03 2018
 """
 
 from .AbovegroundCompetition import AbovegroundCompetition
+from .SimpleTest import SimpleTest
