+++ TreeModelLib/BelowgroundCompetition/BelowgroundCompetition.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+from TreeModelLib import TreeModel
 
 
-class BelowgroundCompetition:
+class BelowgroundCompetition(TreeModel):
     ## Concept for belowground competition
     #  @VAR case: belowground competition concept to be used for the model.
     #  @date: 2019 - Today
@@ -20,9 +21,6 @@ class BelowgroundCompetition:
         from .SimpleTest import SimpleTest
         self.concept = SimpleTest(args)
 
-    def addTree(self, x, y, geometry, parameter):
-        self.concept.addTree(x, y, geometry, parameter)
-
     def getBelowgroundResources(self):
         return self.belowground_resources
 
@@ -30,5 +28,5 @@ class BelowgroundCompetition:
         self.concept.calculateBelowgroundResources()
         self.belowground_resources = self.concept.getBelowgroundResources()
 
-    def prepareNextTimeStep(self, t_ini, t_end):
-        self.concept.prepareNextTimeStep(t_ini, t_end)
+    def getConceptType(self):
+        return "belowground competition concept"
