+++ TreeModelLib/__init__.py
@@ -5,7 +5,7 @@ Created on Thu Nov  8 15:25:03 2018
 
 @author: bathmann
 """
-
+from .TreeModel import TreeModel
 from .AbovegroundCompetition import AbovegroundCompetition
 from .BelowgroundCompetition import BelowgroundCompetition
 from .GrowthAndDeathDynamics import GrowthAndDeathDynamics
